Examining Blinking Behavior of single quantum dot blinking
The main purpose of this repository is to describe a little bit of quantum dot blinking followed with a description of how to use these programs
for other uses.

Blinking is defined as the process randomly switching between high light emitting and low light emitting states. One way to think about this
is a random walk process where at one moment you might be in a bright state and can continue for a long period of time and the next it
can be in a dim state. This was a tricky problem to solve since it follows a different kinetic behavior than seen in a common general 
chemistry course. It follows what is known as power law statisitics or the probability density of the on and off states follow
p(t) = At^-m where p is the probability density, t is time for on or off states, A is a constant and m is the characterisitc power law behavior.

The first program i wrote goes about solving fitting the raw data to a power law plot.